Continue to port appli on windows (compil OK under minGW)

+++ src/common/libmmgf.h
@@ -11,7 +11,7 @@
 ! ** See the "libmmg*.h" file for a correct displaying of the documentation.
 ! */
 
-#define MMG5_DATA_PTR_T INTEGER(kind=8)
+#define MMG5_DATA_PTR_T INTEGER(kind=4)
 ! /* =============================================================================
 ! **  This file is part of the mmg software package for the tetrahedral
 ! **  mesh modification.
@@ -43,6 +43,8 @@
 !  * \version 5
 !  * \date 01 2014
 !  * \copyright GNU Lesser General Public License.
+!  * \warning To keep the genheader working, don't break line between the enum
+!  * name and the opening brace (it creates errors under windows)
 !  */
